Initial compliance test passed
